+++ components/badge/__tests__/index.test.js
@@ -128,6 +128,7 @@ describe('Badge', () => {
   });
 
   // https://github.com/ant-design/ant-design/issues/21331
+  // https://github.com/ant-design/ant-design/issues/31590
   it('render Badge status/color when contains children', () => {
     const wrapper = mount(
       <div>
@@ -143,6 +144,9 @@ describe('Badge', () => {
       </div>,
     );
     expect(wrapper.render()).toMatchSnapshot();
+    expect(wrapper.find(Badge).at(0).find('.ant-scroll-number-only-unit').text()).toBe('5');
+    expect(wrapper.find(Badge).at(1).find('.ant-scroll-number-only-unit').text()).toBe('5');
+    expect(wrapper.find(Badge).at(2).find('.ant-scroll-number-only-unit').text()).toBe('5');
   });
 
   it('Badge should work when status/color is empty string', () => {
@@ -154,74 +158,5 @@ describe('Badge', () => {
     );
 
     expect(wrapper.find('.ant-badge')).toHaveLength(2);
-  });
-});
-
-describe('Ribbon', () => {
-  mountTest(Badge.Ribbon);
-  rtlTest(Badge.Ribbon);
-
-  describe('placement', () => {
-    it('works with `start` & `end` placement', () => {
-      const wrapperStart = mount(
-        <Badge.Ribbon placement="start">
-          <div />
-        </Badge.Ribbon>,
-      );
-      expect(wrapperStart.find('.ant-ribbon-placement-start').length).toEqual(1);
-
-      const wrapperEnd = mount(
-        <Badge.Ribbon placement="end">
-          <div />
-        </Badge.Ribbon>,
-      );
-      expect(wrapperEnd.find('.ant-ribbon-placement-end').length).toEqual(1);
-    });
-  });
-
-  describe('color', () => {
-    it('works with preset color', () => {
-      const wrapper = mount(
-        <Badge.Ribbon color="green">
-          <div />
-        </Badge.Ribbon>,
-      );
-      expect(wrapper.find('.ant-ribbon-color-green').length).toEqual(1);
-    });
-    it('works with custom color', () => {
-      const wrapperLeft = mount(
-        <Badge.Ribbon color="#888" placement="start">
-          <div />
-        </Badge.Ribbon>,
-      );
-      expect(wrapperLeft.find('.ant-ribbon').prop('style').background).toEqual('#888');
-      expect(wrapperLeft.find('.ant-ribbon-corner').prop('style').color).toEqual('#888');
-      const wrapperRight = mount(
-        <Badge.Ribbon color="#888" placement="end">
-          <div />
-        </Badge.Ribbon>,
-      );
-      expect(wrapperRight.find('.ant-ribbon').prop('style').background).toEqual('#888');
-      expect(wrapperRight.find('.ant-ribbon-corner').prop('style').color).toEqual('#888');
-    });
-  });
-
-  describe('text', () => {
-    it('works with string', () => {
-      const wrapper = mount(
-        <Badge.Ribbon text="cool">
-          <div />
-        </Badge.Ribbon>,
-      );
-      expect(wrapper.find('.ant-ribbon').text()).toEqual('cool');
-    });
-    it('works with element', () => {
-      const wrapper = mount(
-        <Badge.Ribbon text={<span className="cool" />}>
-          <div />
-        </Badge.Ribbon>,
-      );
-      expect(wrapper.find('.cool').length).toEqual(1);
-    });
   });
 });
