+++ components/badge/demo/status.md
@@ -1,5 +1,5 @@
 ---
-order: 6
+order: 8
 title:
   zh-CN: 状态点
   en-US: Status
@@ -17,7 +17,7 @@ Standalone badge with status.
 import { Badge } from 'antd';
 
 ReactDOM.render(
-  <div>
+  <>
     <Badge status="success" />
     <Badge status="error" />
     <Badge status="default" />
@@ -33,7 +33,7 @@ ReactDOM.render(
     <Badge status="processing" text="Processing" />
     <br />
     <Badge status="warning" text="Warning" />
-  </div>,
+  </>,
   mountNode,
 );
 ```
